fix(takserver): address CodeRabbit review + regenerate stale screenshot refs
CodeRabbit's second pass on PR #6554:

- TAKServerManager.stop() and the failed-start branch cleared onMessage but
  left onClientConnected registered. A stray callback after stop() or a
  failed start() would still emit into clientConnected, which a later
  start() could observe. Both callbacks are now cleared together.
- MeshNodeCoTConversion's hand-rolled formatOneDecimal is replaced with the
  shared NumberFormatter.format(value, 1) — core:takserver already depends
  on core:common, and NumberFormatter also handles NaN/infinite inputs the
  local version didn't.
- Deduplicated the two independently hand-written FakeTAKServerManager
  fakes (TAKMeshIntegrationTest, MeshToCotBroadcasterTest) into one shared
  FakeTAKServerManager.kt in this module's commonTest.
- Deduplicated TAKConfigPermissionDeniedTest's local FakeTakPrefs in favor
  of the shared core:testing one (added the module dependency); its two
  tests that relied on the local fake defaulting isTakServerEnabled to true
  now seed that explicitly, since the shared fake defaults to false.

Not adopted: extracting FakeTAKServerManager into :core:testing, as
suggested. core:testing's own build.gradle.kts documents an explicit
boundary — "Heavy modules ... should depend on core:testing, not vice
versa" — and TAKServerManager lives in core:takserver, one of those heavy
modules. Deduplicating within core:takserver's own commonTest gets the
same benefit without the dependency reversal.

Not adopted: sharing one TestCoroutineScheduler between MeshToCotBroadcaster's
CoroutineDispatchers and the harness's runTest(UnconfinedTestDispatcher()) in
TAKMeshIntegrationTest / MeshServiceOrchestratorTest. Correctly flagged as a
latent trap for a future test that enables the mesh-to-CoT pref without
fixing this, but today it's inert (the pref stays disabled, so the
broadcaster's delay()-based coroutines never run), and fixing it properly
means restructuring all 10 runTest call sites in TAKMeshIntegrationTest to
thread a shared dispatcher through — out of proportion to a trivial,
currently-dormant nitpick.

Also regenerated the two screenshot references CI flagged as stale
(ScreenshotTakServerSectionEnabled Light/Dark) — expected fallout of the
new SwitchPreference row added to TakServerSection in the prior commit.

Refs #6416
